+++ src/main/java/com/twinsession/mixin/PlayerList_TwinSessionMixin.java
@@ -1,5 +1,6 @@
 package com.twinsession.mixin;
 
+import com.mojang.authlib.GameProfile;
 import com.twinsession.TwinSession;
 import net.minecraft.network.Connection;
 import net.minecraft.server.level.ServerPlayer;
@@ -11,12 +12,10 @@ import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
-import java.util.UUID;
-
 @Mixin(PlayerList.class)
 public class PlayerList_TwinSessionMixin {
     @Inject(method = "disconnectAllPlayersWithProfile", at = @At("HEAD"), cancellable = true)
-    private void onDisconnectAllPlayersWithProfile(UUID uUID, CallbackInfoReturnable<Boolean> cir) {
+    private void onDisconnectAllPlayersWithProfile(GameProfile gameProfile, CallbackInfoReturnable<Boolean> cir) {
         // Prevent disconnection of existing players
         cir.setReturnValue(false);
     }
@@ -24,6 +23,12 @@ public class PlayerList_TwinSessionMixin {
     @Inject(method = "placeNewPlayer", at = @At("TAIL"))
     private void afterPlaceNewPlayer(Connection clientConnection, ServerPlayer playerIn, CommonListenerCookie cookie, CallbackInfo ci) {
         TwinSession.playerJoined(playerIn);
+    }
+
+    @Inject(method = "placeNewPlayer", at = @At(value = "INVOKE", target = "Lnet/minecraft/server/network/ServerGamePacketListenerImpl;send(Lnet/minecraft/network/protocol/Packet;)V", ordinal = 0))
+    private void injectBeforePlayerInfoUpdate(Connection connection, ServerPlayer serverPlayer, CommonListenerCookie commonListenerCookie, CallbackInfo ci) {
+        // Have to set before joining, otherwise it will not render correctly.
+        TwinSession.copySourceTexture(serverPlayer);
     }
 
     @Inject(method = "remove", at = @At("TAIL"))